Back_End_Netflix/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
pillow
django-rest-authtoken